updated bdd for proper import refactor
